Change Array.clone to return iso^
The signature for Array.clone was:
	fun clone(): Array[this->A!]^ =>
This commit changes it to:
	fun clone(): Array[this->A!] iso^ =>

This brings Array.clone in line with the semantics of String.clone.

This is important to be able to easily make copies of local Arrays
that can be sent to other actors. That is, given an actor, Holder
can now easily construct a copy of an Array to send:

actor Holder
  let _collected: Array[T val] iso
  ...
  be send() =>
    destination.process(_collected.clone())

Previously this might have been done in the following contrived way:
actor Holder
  let _collected: Array[T val] iso
  ...
  be send() =>
    destination.process(_cpy())

  fun ref _cpy(): Array[T val] val =>
    recover val
      let c: Array[T val] ref = Array[T val](0)
      let s = _collected.size()
      var i: USize = 0
      while i < s do
        try
          c.push(_collected(i))
        end
        i = i + 1
      end
      c
    end

This change also makes it possible to send copies when the local Array
is a ref field vs iso field.